alles andersch

+++ DigitIsolation.py
@@ -74,7 +74,7 @@ class DigitIsolation:
             [0, 0],
             [maxWidth - 1, 0],
             [maxWidth - 1, maxHeight - 1],
-            [0, maxHeight - 1]], dtype="float32")
+            [0, maxHeight - 1]], dtype="int8")
 
         # compute the perspective transform matrix and then apply it
         M = cv2.getPerspectiveTransform(rect, dst)
